Split ComponentUtilities into traits
