+++ src/Traits/ComponentUtilities.php
@@ -7,27 +7,21 @@ use Livewire\Attributes\Locked;
 use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\ComponentConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\ComponentHelpers;
+use Rappasoft\LaravelLivewireTables\Traits\Core\Component\{HandlesComputedProperties,HandlesEmptyMessage, HandlesFingerprint, HandlesOfflineIndicator,HandlesTableName};
 
 trait ComponentUtilities
 {
-    use ComponentConfiguration,
+    use HandlesTableName,
+        HandlesFingerprint,
+        HandlesEmptyMessage,
+        HandlesComputedProperties,
+        HandlesOfflineIndicator,
+        ComponentConfiguration,
         ComponentHelpers;
-
+        
     public array $table = [];
 
     protected $model;
-
-    #[Locked]
-    public string $tableName = 'table';
-
-    #[Locked]
-    public ?string $dataTableFingerprint;
-
-    protected bool $offlineIndicatorStatus = true;
-
-    protected string $emptyMessage = 'No items found, try to broaden your search';
-
-    protected bool $useComputedProperties = true;
 
     protected bool $hasRunConfigure = false;
 
@@ -80,17 +74,6 @@ trait ComponentUtilities
             $this->hasRunConfigure = true;
 
         }
-    }
-
-    /**
-     * Returns a unique id for the table, used as an alias to identify one table from another session and query string to prevent conflicts
-     */
-    protected function generateDataTableFingerprint(): string
-    {
-        $className = str_split(static::class);
-        $crc32 = sprintf('%u', crc32(serialize($className)));
-
-        return base_convert($crc32, 10, 36);
     }
 
     /**
